DEV/adapters/requirements-adapters.txt
aiohttp
aiofiles
aiomqtt
aioredis
aiohttp
aiosmtplib
httpx
pydantic
websockets
python-dotenv
requests
paho-mqtt
fastapi
pyyaml
discord.py
slack_sdk
python-telegram-bot
asyncio
asyncio-mqtt
socketio
python-socketio[asyncio]
PyJWT
cryptography
oauth2lib
jsonschema
pytest
pytest-asyncio
pytest-mock
pytest-cov
black
flake8
mypy
pre-commit
zabbix-api
homeassistant-api
google-auth
google-auth-oauthlib
google-auth-httplib2
python-homey
structlog
prometheus-client
configparser
uvicorn
uvloop
ujson
tenacity
scikit-learn
numpy
pandas
matplotlib
seaborn
pyjwt
redis
